whatpm/t/xml/cdata-1.dat

#data
a<![CDATA[bb]]>ccc
#errors
1;1;text outside of root element;;;m
1;18;no root element;;;m
#document
| "abbccc"
